+++ tests/python/codegen/test_target_codegen_blob.py
@@ -22,7 +22,7 @@ import numpy as np
 
 import tvm
 import tvm.testing
-from tvm.contrib import cc, popen_pool, tar, utils
+from tvm.support import cc, popen_pool, tar, utils
 from tvm.script import ir as I
 from tvm.script import tirx as T
