+++ tests/python/runtime/test_runtime_rpc.py
@@ -31,7 +31,7 @@ import tvm_ffi
 import tvm
 import tvm.testing
 from tvm import rpc, te
-from tvm.contrib import cc, utils
+from tvm.support import cc, utils
 from tvm.rpc.proxy import Proxy
 from tvm.rpc.tracker import Tracker
 from tvm.script import ir as I
